requirements.txt
torch
pandas
scikit-learn
numpy
python-dotenv